requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
colorama==0.4.4
discord.py==1.7.2
idna==3.1
loguru==0.5.3
multidict==5.1.0
typing-extensions==3.10.0.0
win32-setctime==1.0.3
yarl==1.6.3